requirements.txt
# Essential dependencies for Label Studio + YOLO workflow
google-cloud-storage>=2.10.0  # For downloading from private GCS buckets
ultralytics>=8.0.0           # YOLO model inference
label-studio-sdk>=0.0.34     # Label Studio API integration
opencv-python>=4.8.0         # Image processing
numpy>=1.24.0               # Numerical computations
Pillow>=10.0.0              # Image handling
PyYAML>=6.0                 # Configuration files
requests>=2.31.0            # HTTP requests
tqdm>=4.65.0                # Progress bars 